% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anota2seqAllGenerics.R, R/anota2seqMethods.R
\docType{methods}
\name{anota2seq.get.contrasts}
\alias{anota2seq.get.contrasts}
\alias{anota2seq.get.contrasts,anota2seqDataSet-method}
\title{Accessor for the 'contrasts' slot of an anota2seqDataSet object.}
\usage{
anota2seq.get.contrasts(object = NULL)

\S4method{anota2seq.get.contrasts}{anota2seqDataSet}(object = NULL)
}
\arguments{
\item{object}{A anota2seqDataSet.}
}
\value{
A matrix containing information about the analyzed contrasts to be 
  used for the analysis.
}
\description{
Retrieve from an anota2seqDataSet the contrast matrix that is used in the
analysis
}
\examples{
data(anota2seqData)
# Initialize anota2seqDataSet
anota2seqDataSet <- anota2seqDataSetFromMatrix(
    dataP = anota2seqDataP[1:100,],
    dataT = anota2seqDataT[1:100,],
    phenoVec = anota2seqPhenoVec,
    dataType = "RNAseq",
    normalize = TRUE)
# Run analysis of differential translation
anota2seqDataSet <- anota2seqAnalyze(anota2seqDataSet, 
                                     analysis = "translation")
contr <- anota2seq.get.contrasts(anota2seqDataSet)
contr
#           treatment
# ctrl             -1
# treatment         1
}
\seealso{
\code{\link{anota2seqDataSet-class}}
}
