% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anota2seqAllGenerics.R, R/anota2seqMethods.R
\docType{methods}
\name{anota2seq.get.covariates}
\alias{anota2seq.get.covariates}
\alias{anota2seq.get.covariates,anota2seqDataSet-method}
\title{Accessor for the 'phenoVec' and 'batchVec' slots of an anota2seqDataSet 
object.}
\usage{
anota2seq.get.covariates(object = NULL)

\S4method{anota2seq.get.covariates}{anota2seqDataSet}(object = NULL)
}
\arguments{
\item{object}{An anota2seqDataSet.}
}
\value{
A list of 2 elements named "phenoVec" and "batchVec"
  containing the corresponding covariates to be used for the analysis.
}
\description{
Retrieve from an anota2seqDataSet the covariate(s) to be used in the APV 
model i.e. the annotation of the samples according to their treatments and
batch (when provided).
}
\examples{
data(anota2seqData)
# Initialize anota2seqDataSet
anota2seqDataSet <- anota2seqDataSetFromMatrix(
    dataP = anota2seqDataP[1:100,],
    dataT = anota2seqDataT[1:100,],
    phenoVec = anota2seqPhenoVec,
    batchVec = c(1, 2, 3, 4, 1, 2, 3, 4),
    dataType = "RNAseq",
    normalize = TRUE)
# Run analysis of differential translation
anota2seqDataSet <- anota2seqAnalyze(anota2seqDataSet, 
                                     analysis = "translation")
covariates <- anota2seq.get.covariates(anota2seqDataSet)
covariates
# $phenoVec
# [1] "ctrl"      "ctrl"      "ctrl"      "ctrl"      "treatment" "treatment"
# [7] "treatment" "treatment"
# 
# $batchVec
# [1] "1" "2" "3" "4" "1" "2" "3" "4"
}
\seealso{
\code{\link{anota2seqDataSet-class}}
}
