% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anota2seqData.R
\docType{data}
\name{anota2seqData}
\alias{anota2seqData}
\alias{anota2seqDataP}
\alias{anota2seqDataT}
\alias{anota2seqPhenoVec}
\title{Sample data set for anota2seq}
\format{Eight samples were simulated from 2 sample classes ("control" and 
  "treatment"); both total mRNA (anota2seqDataT) and paired translated mRNA
  (anota2seqDataP) are provided together with a sample class vector
  (anota2seqPhenoVec).
  
  anota2seqDataT and anota2seqDataP are 2 matrices with 10736 rows and 8 
  columns simulated to represent raw RNAseq counts from 8 samples (4 
  replicates of each condition) and 10736 mRNAs. Columns are ordered 
  according to the RNA type pairing i.e. column 1 of anota2seqDataT is from 
  the same sample as column 1 of anota2seqDataP etc.}
\source{
These data originate from the study by \cite{Oertlin et al.} which 
  compared methods for analysis of translatomes quantified by RNAseq.
}
\usage{
data(anota2seqData)
}
\description{
Simulated data used to illustrate the anota2seq workflow.
}
\examples{
##load data set 
data(anota2seqData)
##check dimensions
dim(anota2seqDataP)
head(anota2seqDataP)
dim(anota2seqDataT)
head(anota2seqDataT)
anota2seqPhenoVec
  
}
\references{
Oertlin C. et al. Genome-wide analysis of differential 
  translation and differential translational buffering using anota2seq, 
  bioRxiv, 2017.
}
\keyword{datasets}
