% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anota2seqShow.R
\docType{methods}
\name{show,anota2seqDataSet-method}
\alias{show,anota2seqDataSet-method}
\alias{show}
\title{Show information for an anota2seqDataSet}
\usage{
\S4method{show}{anota2seqDataSet}(object = NULL)
}
\arguments{
\item{object}{An anota2seqDataSet.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function gives basic information on the slots within your
anota2seqDataSet.
}
\examples{
data(anota2seqData)
anota2seqDataSet <- anota2seqDataSetFromMatrix(dataP = anota2seqDataP[1:1000,],
                                      dataT = anota2seqDataT[1:1000,],
                                      phenoVec = anota2seqPhenoVec,
                                      dataType = "RNAseq",
                                      normalize = TRUE)
                                      
show(anota2seqDataSet)
}
