% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anota2seqAllGenerics.R, R/anota2seqMethods.R
\docType{methods}
\name{anota2seq.get.normalizedData}
\alias{anota2seq.get.normalizedData}
\alias{anota2seq.get.normalizedData,anota2seqDataSet-method}
\title{Accessor for the 'normalizedDataP' and 'normalizedDataT' slots of an anota2seqDataSet object.}
\usage{
anota2seq.get.normalizedData(object = NULL)

\S4method{anota2seq.get.normalizedData}{anota2seqDataSet}(object = NULL)
}
\arguments{
\item{object}{A anota2seqDataSet.}
}
\value{
A list of 2 elements named "dataP" and "dataT"
  containing the corresponding normalized data to be used for the analysis.
}
\description{
Retrieve the data that is used in the analysis (i.e. if raw RNAseq counts 
were supplied, this function retrieves the normalized counts used in the 
analysis).
}
\examples{
data(anota2seqData)
# Initialize anota2seqDataSet
anota2seqDataSet <- anota2seqDataSetFromMatrix(
    dataP = anota2seqDataP[1:1000,],
    dataT = anota2seqDataT[1:1000,],
    phenoVec = anota2seqPhenoVec,
    dataType = "RNAseq",
    normalize = TRUE)
# Get normalized dataP
normDataP <- anota2seq.get.normalizedData(anota2seqDataSet)[["dataP"]]
# Get normalized dataT
normDataT <- anota2seq.get.normalizedData(anota2seqDataSet)[["dataT"]]

}
\seealso{
\code{\link{anota2seqDataSet-class}}
}
