% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anota2seqShow.R
\docType{methods}
\name{show,Anota2seqDataSet-method}
\alias{show,Anota2seqDataSet-method}
\alias{show}
\title{Show information for an Anota2seqDataSet}
\usage{
\S4method{show}{Anota2seqDataSet}(object)
}
\arguments{
\item{object}{An Anota2seqDataSet.}
}
\value{
No value is returned; this function is called for its side effects.
}
\description{
This function gives basic information on the slots within your
Anota2seqDataSet.
}
\examples{
data(anota2seq_data)
Anota2seqDataSet <- anota2seqDataSetFromMatrix(dataP = anota2seq_data_P[1:1000,],
                                      dataT = anota2seq_data_T[1:1000,],
                                      phenoVec = anota2seq_pheno_vec,
                                      dataType = "RNAseq",
                                      normalize = TRUE)
                                      
show(Anota2seqDataSet)
}
